+++ cudaO-implementation/utils/utils.h
@@ -9,6 +9,6 @@ __device__ uint8_t getBit(uint8_t* data, int& position);
 __device__ uint16_t getNBits(uint8_t* data, int& position, int n);
 __device__ uint8_t getByte(uint8_t* data, int& position);
 __device__ void getNBytes(uint8_t* arr, int length, uint8_t* data, int& position);
-__device__ int decodeNumber(uint8_t code, int bits);
+__device__ int16_t decodeNumber(uint8_t code, int bits);
 
 #endif
